5.6.1 Поправит библиотеку @project/blog-likes.
