+++ project/libs/blog/likes/src/module/likes.module.ts
@@ -4,10 +4,11 @@ import { LikesController } from './likes.controller';
 import { LikesRepository } from './likes.repository';
 import { PrismaClientModule } from '@project/blog-models';
 import { LikeFactory } from './likes.factory';
+import { LikesTechController } from './likes-tech.controller';
 
 @Module({
   imports: [PrismaClientModule],
-  controllers: [LikesController],
+  controllers: [LikesController, LikesTechController],
   providers: [LikesService, LikesRepository, LikeFactory],
   exports: [LikesService],
 })
